feat: 캘린더 주간 네비+오늘 이동을 하단 중앙 플로팅으로 통합
상단 고정 CalendarRangeHeader(날짜범위/주이동)와 우하단 TodayButton을
하단 중앙 플로팅 nav 하나로 합쳤다(FloatingCalendarNav). 날짜 범위·주
이동·오늘로 이동을 한 컨트롤에서 제공한다. 두 구컴포넌트는 제거.
